requirements.txt
asgiref==3.5.2
astroid==2.11.5
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
dill==0.3.5.1
dj-database-url==0.5.0
Django==4.0.5
django-cleanup==6.0.0
django-filter==21.1
django-phonenumber-field==6.3.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
phonenumberslite==8.12.50
Pillow==9.1.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pylint==2.14.1
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
telesign==2.2.2
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
